Module Linked_set.Make

Functorized lock-free linked set with length for testing and as an example of using Size. The functorization is to allow the use of traced atomics with DSCheck.

Parameters

module Atomic : sig ... end

module Size : sig ... end

Signature

type 'a t

val create : unit -> 'a t

val length : 'a t -> int

val mem : 'a t -> 'a -> bool

val try_add : 'a t -> 'a -> bool

val try_remove : 'a t -> 'a -> bool