Module Not_lockfree.CAS_interface

Alternative interface, which may perform better on architectures without FAD instructions (e.g. AArch).

CAS_interface should not be the default choice. It may be a little faster on ARM, but it is going to be a few times slower than standard on x86.

val push : 'a t -> 'a -> bool

val pop : 'a t -> 'a option